Compute Simultaneous Confidence Bands for Event Study Coefficients in Functional DiD Framework

The fdid_scb function is used to compute simultaneous confidence bands for event study coefficients in a functional framework. Specifically, we compute an infimum-based simultaneous confidence band in pre-treatment period by parametric bootstrap, and a supremum-based simultaneous confidence band in post-treatment period by the algorithm of Kac-Rice formula proposed in Liebl and Reimherr (2023).

Usage

fdid_scb(
  object = NULL,
  beta = NULL,
  cov = NULL,
  t0 = NULL,
  df = NULL,
  ci.alpha = 0.05,
  scb.pre.alpha = 0.05,
  scb.post.alpha = 0.05
)

Arguments

object

an object of S3 class "fdid", which is an output of the function fdid. If it is provided, arguments 'beta', 'cov' and 't0' are ignored.

beta

a numeric matrix with two columns. The first column is the estimates of event study coefficients and the second column is the corresponding event time. The estimate of coefficient at reference time should be normalized to 0.

cov

a numeric matrix of covariance estimates.

t0

a numeric scalar that indicates the reference time.

df

degree of freedom for the t-distributed based band in post-treatment periods. If NULL, Gaussian distributed bands are computed.

ci.alpha

significance level for the point-wise confidence intervals.

scb.pre.alpha

significance level for the infimum-based simultaneous confidence band in pre-treatment periods.

scb.post.alpha

significance level for the supremum-based simultaneous confidence band in post-treatment periods.

Value

The fdid_scb function returns a list containing three objects:

scb

a list that includes the functional estimate of event study coefficients and the simultaneous confidence bands.

ci

a matrix that includes the point-wise estimates of event study coefficients and their confidence intervals.

data

a list that stores the original inputs of beta and cov.

The output is an object of S3 class "fdid_scb".

References

Fang, C. and Liebl, D. (2025). Making Event Study Plots Honest: A Functional Data Approach to Causal Inference. arXiv:2512.06804.

Liebl, D. and M. Reimherr (2023). Fast and fair simultaneous confidence bands for functional parameters. Journal of the Royal Statistical Society Series B: Statistical Methodology 85(3), 842–868

See also

Examples

data(LWdata)
fdid_scb_est1 <- fdid_scb(beta=LWdata$beta, cov=LWdata$cov, t0=LWdata$t0)

data(simulated_stagger_example)
fdid_est2 <- fdid(data=simulated_stagger_data, treatment=simulated_stagger_treatment)
fdid_scb_est2 <- fdid_scb(object=fdid_est2)